requirements.txt
PyQt6==6.11.0
python-barcode[images]==0.16.1
Pillow==11.3.0
pandas==3.0.2
openpyxl==3.1.5
